format spring-3.0.x-scenario
