+++ test/plugin/scenarios/spring-3.0.x-scenario/src/main/java/test/apache/skywalking/apm/testcase/spring3/CaseController.java
@@ -16,13 +16,13 @@
  *
  */
 
-package test.org.apache.skywalking.apm.testcase.spring3;
+package test.apache.skywalking.apm.testcase.spring3;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-import test.org.apache.skywalking.apm.testcase.spring3.service.TestServiceBean;
+import test.apache.skywalking.apm.testcase.spring3.service.TestServiceBean;
 
 @Controller
 public class CaseController {
