+++ test/plugin/scenarios/spring-3.0.x-scenario/src/main/java/test/apache/skywalking/apm/testcase/spring3/dao/TestRepositoryBean.java
@@ -16,7 +16,7 @@
  *
  */
 
-package test.org.apache.skywalking.apm.testcase.spring3.dao;
+package test.apache.skywalking.apm.testcase.spring3.dao;
 
 import org.springframework.stereotype.Repository;
